pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "ds2dbx"
version = "0.1.0"
description = "CLI tool for end-to-end IBM DataStage to Databricks migration using Lakebridge"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
ds2dbx = "ds2dbx.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
